Revert "refactor(Select): subcomponents structure"
This reverts commit 502c20d80b9ce93a1db236fab9998aaf770ef7b8.
