+++ packages/react-ui-kit/src/Form/Select/BaseSelect/BaseSelectMenuList.tsx
@@ -19,13 +19,13 @@
 
 import {components, MenuListProps} from 'react-select';
 
-import {closeButtonStyles, headingContainerStyles} from './SelectMenuList.styles';
+import {closeButtonStyles, headingContainerStyles} from './BaseSelectMenuList.styles';
 
 import {CloseIcon} from '../../../Icon';
 import {Theme} from '../../../Layout';
 
 // eslint-disable-next-line react/display-name
-export const SelectMenuList = (menuListHeading: string, dataUieName: string) => (props: MenuListProps) => {
+export const BaseSelectMenuList = (menuListHeading: string, dataUieName: string) => (props: MenuListProps) => {
   const {selectProps, children} = props;
 
   const handleClose = () => {
@@ -52,4 +52,4 @@ export const SelectMenuList = (menuListHeading: string, dataUieName: string) => 
     </components.MenuList>
   );
 };
-SelectMenuList.displayName = 'SelectMenuList';
+BaseSelectMenuList.displayName = 'BaseSelectMenuList';
