+++ packages/react-ui-kit/src/Form/Select/ComboboxSelect/ComboboxSelect.stories.tsx
@@ -25,6 +25,10 @@ const options = [
   {value: '1', label: 'Option 1'},
   {value: '2', label: 'Option 2'},
   {value: '3', label: 'Option 3'},
+  {value: '4', label: 'Option 4'},
+  {value: '5', label: 'Option 5'},
+  {value: '6', label: 'Option 6'},
+  {value: '7', label: 'Option 7'},
 ];
 
 const meta = {
@@ -52,9 +56,6 @@ export const Default: Story = {
     options,
     placeholder: 'Select options...',
     dataUieName: 'default-select',
-    createOptionLabel: inputValue => `Create item "${inputValue}"`,
-    onCreateOption: () => {},
-    noOptionsMessage: 'No options available',
   },
 };
 
@@ -65,9 +66,6 @@ export const WithValue: Story = {
     value: [options[0], options[1]],
     placeholder: 'Select options...',
     dataUieName: 'with-value-select',
-    createOptionLabel: inputValue => `Create item "${inputValue}"`,
-    onCreateOption: () => {},
-    noOptionsMessage: 'No options available',
   },
 };
 
@@ -78,9 +76,6 @@ export const Disabled: Story = {
     isDisabled: true,
     placeholder: 'Select options...',
     dataUieName: 'disabled-select',
-    createOptionLabel: inputValue => `Create item "${inputValue}"`,
-    onCreateOption: () => {},
-    noOptionsMessage: 'No options available',
   },
 };
 
@@ -90,10 +85,6 @@ export const Creatable: Story = {
     options,
     placeholder: 'Select or create options...',
     dataUieName: 'creatable-select',
-    createOptionLabel: inputValue => `Create item "${inputValue}"`,
-    onCreateOption: () => {},
-    noOptionsMessage: 'No options available',
-    required: true,
   },
 };
 
@@ -102,8 +93,5 @@ export const WithLabel: Story = {
     id: 'with-label-select',
     label: 'Select options',
     options,
-    createOptionLabel: inputValue => `Create item "${inputValue}"`,
-    onCreateOption: () => {},
-    noOptionsMessage: 'No options available',
   },
 };
