+++ packages/react-ui-kit/src/Form/Select/ComboboxSelect/ComboboxSelect.tsx
@@ -18,12 +18,10 @@
  */
 
 import {useTheme} from '@emotion/react';
-import {components, MultiValueRemoveProps, NoticeProps} from 'react-select';
 import CreatableSelect from 'react-select/creatable';
 
-import {customStyles, noOptionsMessageStyles, wrapperStyles} from './ComboboxSelect.styles';
+import {customStyles} from './ComboboxSelect.styles';
 
-import {CloseIcon} from '../../../Icon/CloseIcon';
 import {Theme} from '../../../Layout';
 import {InputLabel} from '../../InputLabel';
 import {BaseSelectDropdownIndicator} from '../BaseSelect/BaseSelectDropdownIndicator';
@@ -37,9 +35,7 @@ export interface ComboboxSelectProps {
   isDisabled?: boolean;
   placeholder?: string;
   dataUieName?: string;
-  onCreateOption: (inputValue: string) => void;
-  createOptionLabel: (inputValue: string) => string;
-  noOptionsMessage: string;
+  onCreateOption?: (inputValue: string) => void;
   label?: string;
   required?: boolean;
 }
@@ -53,15 +49,25 @@ export const ComboboxSelect = ({
   placeholder,
   dataUieName,
   onCreateOption,
-  createOptionLabel,
-  noOptionsMessage,
   label,
   required,
 }: ComboboxSelectProps) => {
   const theme = useTheme() as Theme;
 
   return (
-    <div css={wrapperStyles} data-uie-name={dataUieName}>
+    <div
+      css={{
+        marginBottom: '20px',
+        width: '100%',
+        position: 'relative',
+        '& .select__menu': {
+          position: 'absolute',
+          width: '100%',
+          zIndex: 1,
+        },
+      }}
+      data-uie-name={dataUieName}
+    >
       {label && (
         <InputLabel htmlFor={id} isRequired={required}>
           {label}
@@ -78,28 +84,14 @@ export const ComboboxSelect = ({
         placeholder={placeholder}
         styles={customStyles({theme})}
         classNamePrefix="select"
-        formatCreateLabel={createOptionLabel}
+        formatCreateLabel={inputValue => `Create item "${inputValue}"`}
         onCreateOption={onCreateOption}
         closeMenuOnSelect={false}
         components={{
           ClearIndicator: () => null,
           DropdownIndicator: BaseSelectDropdownIndicator,
-          MultiValueRemove: props => <MultiValueRemove {...props} theme={theme} />,
-          NoOptionsMessage: props => <NoOptionsMessage {...props} message={noOptionsMessage} />,
         }}
       />
     </div>
   );
 };
-
-const MultiValueRemove = ({theme, ...props}: MultiValueRemoveProps & {theme: Theme}) => (
-  <components.MultiValueRemove {...props}>
-    <CloseIcon width={10} height={10} />
-  </components.MultiValueRemove>
-);
-
-const NoOptionsMessage = ({message, ...props}: NoticeProps & {message: string}) => (
-  <components.NoOptionsMessage {...props}>
-    <div css={noOptionsMessageStyles}>{message}</div>
-  </components.NoOptionsMessage>
-);
