+++ packages/react-ui-kit/src/Form/Select/Select.tsx
@@ -23,14 +23,14 @@ import {CSSObject, useTheme} from '@emotion/react';
 import ReactSelect, {GroupBase, StylesConfig} from 'react-select';
 import type {StateManagerProps} from 'react-select/dist/declarations/src/stateManager';
 
+import {BaseSelectContainer} from './BaseSelect/BaseSelectContainer';
 import {BaseSelectDropdownIndicator} from './BaseSelect/BaseSelectDropdownIndicator';
+import {BaseSelectIndicatorsContainer} from './BaseSelect/BaseSelectIndicatorsContainer';
+import {BaseSelectMenu} from './BaseSelect/BaseSelectMenu';
+import {BaseSelectMenuList} from './BaseSelect/BaseSelectMenuList';
+import {BaseSelectOption, isGroup} from './BaseSelect/BaseSelectOption';
+import {BaseSelectValueContainer} from './BaseSelect/BaseSelectValueContainer';
 import {customStyles} from './Select.styles';
-import {SelectContainer} from './SelectContainer/SelectContainer';
-import {SelectIndicatorsContainer} from './SelectIndicatorsContainer/SelectIndicatorsContainer';
-import {SelectMenu} from './SelectMenu/SelectMenu';
-import {SelectMenuList} from './SelectMenuList/SelectMenuList';
-import {isGroup, SelectOption} from './SelectOption/SelectOption';
-import {SelectValueContainer} from './SelectValueContainer/SelectValueContainer';
 
 import {Theme} from '../../Layout';
 import {TabIndex} from '../../types/enums';
@@ -119,13 +119,13 @@ export const Select = <IsMulti extends boolean = false, Group extends GroupBase<
           }) as StylesConfig
         }
         components={{
-          SelectContainer: SelectContainer,
+          SelectContainer: BaseSelectContainer,
           DropdownIndicator: BaseSelectDropdownIndicator,
-          Option: SelectOption(dataUieName),
-          Menu: SelectMenu(dataUieName, menuCSS),
-          ValueContainer: SelectValueContainer,
-          IndicatorsContainer: SelectIndicatorsContainer,
-          ...(menuListHeading && {MenuList: SelectMenuList(menuListHeading, dataUieName)}),
+          Option: BaseSelectOption(dataUieName),
+          Menu: BaseSelectMenu(dataUieName, menuCSS),
+          ValueContainer: BaseSelectValueContainer,
+          IndicatorsContainer: BaseSelectIndicatorsContainer,
+          ...(menuListHeading && {MenuList: BaseSelectMenuList(menuListHeading, dataUieName)}),
         }}
         tabIndex={TabIndex.UNFOCUSABLE}
         isDisabled={disabled}
